requirements.txt
ansible==1.9.4
Babel==2.1.1
cffi==1.3.1
cryptography==1.1.1
d2to1==0.2.11
debtcollector==0.10.0
ecdsa==0.13
enum34==1.0.4
idna==2.0
ipaddress==1.0.15
iso8601==0.1.11
Jinja2==2.8
MarkupSafe==0.23
monotonic==0.4
msgpack-python==0.4.6
ndg-httpsclient==0.4.0
netaddr==0.7.18
netifaces==0.10.4
oslo.config==2.7.0
oslo.i18n==2.7.0
oslo.serialization==1.11.0
oslo.utils==2.8.0
paramiko==1.16.0
pbr==1.8.1
prettytable==0.7.2
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==0.15.1
python-keystoneclient==1.8.1
python-novaclient==2.35.0
pytz==2015.7
PyYAML==3.11
requests==2.8.1
simplejson==3.8.1
six==1.10.0
stevedore==1.9.0
wrapt==1.10.5
